chore: [repeater PR 2] Move nodeSpec tests to nodeSpec.spec.ts (#235)
* Add repeater node to nodeSpec

* Add test to verify NodeSpec creation

* Amend runtime and types to pass test

* Move nodeSpec tests to nodeSpec test file

* Better hierarchy/descriptions in element.spec.ts
