+++ src/plugin/__tests__/elementSpec.spec.ts
@@ -1,10 +1,8 @@
 import { buildElementPlugin } from "../element";
 import { createElementSpec } from "../elementSpec";
-import type { CustomFieldDescription } from "../fieldViews/CustomFieldView";
 import { createNoopElement } from "../helpers/test";
-import { getNodeNameFromField } from "../nodeSpec";
 
-describe("mount", () => {
+describe("createElementSpec", () => {
   describe("fieldView typesafety", () => {
     it("should provide typesafe fieldView to its consumer", () => {
       const fieldDescriptions = {
@@ -87,9 +85,9 @@ describe("mount", () => {
     });
   });
 
-  describe("nodeSpec generation", () => {
+  describe("NodeSpec generation", () => {
     it("should create an nodeSpec with no nodes when the spec is empty", () => {
-      const { nodeSpec } = buildElementPlugin([]);
+      const { nodeSpec } = buildElementPlugin({});
       expect(nodeSpec.size).toBe(0);
     });
 
@@ -102,15 +100,6 @@ describe("mount", () => {
       expect(nodeSpec.get("testElement2")).toMatchObject({ content: "" });
     });
 
-    it("should convert dashes in element names to underscores", () => {
-      const testElement1 = createNoopElement({});
-      const { nodeSpec } = buildElementPlugin({
-        "test-element-1": testElement1,
-      });
-      expect(nodeSpec.size).toBe(1);
-      expect(nodeSpec.get("test_element_1")).toMatchObject({ content: "" });
-    });
-
     it("should add a custom group if specified", () => {
       const testElement1 = createNoopElement({});
       const { nodeSpec } = buildElementPlugin(
@@ -119,149 +108,6 @@ describe("mount", () => {
       );
       expect(nodeSpec.get("testElement1")).toMatchObject({
         group: "customGroup",
-      });
-    });
-
-    it("should create child nodes for each element field, and the parent node should include them in its content expression", () => {
-      const testElement1 = createNoopElement({
-        field1: {
-          type: "richText",
-        },
-        field2: {
-          type: "richText",
-        },
-      });
-      const { nodeSpec } = buildElementPlugin({ testElement1 });
-      expect(nodeSpec.get("testElement1")).toMatchObject({
-        content: "testElement1__field1 testElement1__field2",
-      });
-      expect(nodeSpec.get("testElement1__field1")).toMatchObject({
-        content: "paragraph+",
-      });
-      expect(nodeSpec.get("testElement1__field2")).toMatchObject({
-        content: "paragraph+",
-      });
-    });
-
-    describe("fields", () => {
-      describe("richText", () => {
-        it("should allow the user to specify content, attribute and marks properties", () => {
-          const fieldDescriptions = {
-            field1: {
-              type: "richText" as const,
-              attrs: { customAttr: { default: "custom" } },
-              content: "content",
-              marks: "some marks",
-            },
-          };
-
-          const testElement1 = createNoopElement(fieldDescriptions);
-          const { nodeSpec } = buildElementPlugin({ testElement1 });
-
-          expect(
-            nodeSpec.get(getNodeNameFromField("field1", "testElement1"))
-          ).toMatchObject({
-            content: fieldDescriptions.field1.content,
-            attrs: fieldDescriptions.field1.attrs,
-            marks: fieldDescriptions.field1.marks,
-          });
-        });
-      });
-
-      describe("text", () => {
-        it("should allow the user to specify custom attributes", () => {
-          const fieldDescriptions = {
-            field1: {
-              type: "text" as const,
-              attrs: { customAttr: { default: "custom" } },
-              isMultiline: false,
-              rows: 1,
-              isCode: false,
-            },
-          };
-
-          const testElement1 = createNoopElement(fieldDescriptions);
-          const { nodeSpec } = buildElementPlugin({ testElement1 });
-
-          expect(
-            nodeSpec.get(getNodeNameFromField("field1", "testElement1"))
-          ).toMatchObject({
-            attrs: fieldDescriptions.field1.attrs,
-          });
-        });
-        it("should provide a default inline node spec", () => {
-          const fieldDescriptions = {
-            field1: {
-              type: "text" as const,
-              isMultiline: false,
-              rows: 1,
-              isCode: false,
-            },
-          };
-
-          const testElement1 = createNoopElement(fieldDescriptions);
-          const { nodeSpec } = buildElementPlugin({ testElement1 });
-          const field1NodeSpec = nodeSpec.get(
-            getNodeNameFromField("field1", "testElement1")
-          );
-          expect(field1NodeSpec).toHaveProperty("content", "text*");
-          expect(field1NodeSpec?.parseDOM?.[0]).toMatchObject({
-            tag: "div",
-            preserveWhitespace: false,
-          });
-        });
-      });
-
-      describe("checkbox", () => {
-        it("should specify the appropriate fields on checkbox fields", () => {
-          const fieldDescriptions = {
-            field1: {
-              type: "checkbox" as const,
-              defaultValue: true,
-            },
-          };
-
-          const testElement1 = createNoopElement(fieldDescriptions);
-          const { nodeSpec } = buildElementPlugin({ testElement1 });
-          expect(
-            nodeSpec.get(getNodeNameFromField("field1", "testElement1"))
-          ).toMatchObject({
-            atom: true,
-            attrs: {
-              fields: {
-                default: true,
-              },
-            },
-          });
-        });
-      });
-
-      describe("custom", () => {
-        it("should specify the appropriate fields for custom fields", () => {
-          const fieldDescriptions = {
-            field1: {
-              type: "custom",
-              defaultValue: { arbitraryField: "hai" },
-            } as CustomFieldDescription<{ arbitraryField: string }>,
-          };
-
-          const testElement1 = createNoopElement(fieldDescriptions);
-          const { nodeSpec } = buildElementPlugin({ testElement1 });
-          expect(
-            nodeSpec.get(getNodeNameFromField("field1", "testElement1"))
-          ).toMatchObject({
-            atom: true,
-            attrs: {
-              fields: {
-                default: {
-                  value: {
-                    arbitraryField: "hai",
-                  },
-                },
-              },
-            },
-          });
-        });
       });
     });
   });
